vcs_site/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-js-asset==1.2.2
idna==2.10
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
